Inserter: Fix handling of child blocks
When a block C specifies `parent: [ P ]`, it means that C may only be
added to P. It does NOT mean that P may *only* contain C. (Which is what
`<InnerBlocks allowedBlocks={ [ C ] }>` means.)

This fixes the Inserter so that the correct blocks are shown when
inserting into a block that is referenced by `parent`. It does so by
leaning on `getInserterItems()` which does the right thing.